Introduces meta memory and SWAR-based lookups similar to C++'s `absl::flat_hash_map` hash table (https://abseil.io/docs/cpp/guides/container).
    
Also, reduces `MapOf`'s memory overhead: each bucket now holds up to 5 entries instead of 3.

### Microbenchmarks

A few benchmark results were collected on a laptop with i7-1185G7 and 32GB RAM running Ubuntu 22.04.

Before:
```
BenchmarkMapOfInt_WarmUp/reads=100%-8         	303130814	         4.131 ns/op	 242101267 ops/s	       0 B/op	       0 allocs/op
BenchmarkMapOfInt_WarmUp/reads=99%-8          	232482326	         5.196 ns/op	 192442573 ops/s	       0 B/op	       0 allocs/op
BenchmarkMapOfInt_WarmUp/reads=90%-8          	180387648	         6.811 ns/op	 146829029 ops/s	       0 B/op	       0 allocs/op
BenchmarkMapOfInt_WarmUp/reads=75%-8          	129619245	         9.547 ns/op	 104748466 ops/s	       2 B/op	       0 allocs/op

BenchmarkMapOf_WarmUp/reads=100%-8         	196852258	         6.756 ns/op	 148014399 ops/s	       0 B/op	       0 allocs/op
BenchmarkMapOf_WarmUp/reads=99%-8          	190997551	         6.376 ns/op	 156831274 ops/s	       0 B/op	       0 allocs/op
BenchmarkMapOf_WarmUp/reads=90%-8          	152631080	         8.038 ns/op	 124410942 ops/s	       1 B/op	       0 allocs/op
BenchmarkMapOf_WarmUp/reads=75%-8          	100000000	        11.50 ns/op	  86919108 ops/s	       2 B/op	       0 allocs/op
```

After:
```
BenchmarkMapOfInt_WarmUp/reads=100%-8         	431616265	         2.747 ns/op	 364040402 ops/s	       0 B/op	       0 allocs/op
BenchmarkMapOfInt_WarmUp/reads=99%-8          	320536114	         3.772 ns/op	 265102717 ops/s	       0 B/op	       0 allocs/op
BenchmarkMapOfInt_WarmUp/reads=90%-8          	219056338	         5.281 ns/op	 189374821 ops/s	       0 B/op	       0 allocs/op
BenchmarkMapOfInt_WarmUp/reads=75%-8          	149422306	         7.418 ns/op	 134801273 ops/s	       1 B/op	       0 allocs/op

BenchmarkMapOf_WarmUp/reads=100%-8         	239793300	         4.923 ns/op	 203117147 ops/s	       0 B/op	       0 allocs/op
BenchmarkMapOf_WarmUp/reads=99%-8          	247627736	         4.855 ns/op	 205985869 ops/s	       0 B/op	       0 allocs/op
BenchmarkMapOf_WarmUp/reads=90%-8          	191962411	         6.405 ns/op	 156134268 ops/s	       1 B/op	       0 allocs/op
BenchmarkMapOf_WarmUp/reads=75%-8          	134295975	         9.007 ns/op	 111028357 ops/s	       3 B/op	       0 allocs/op
```